frozen orb sprite and adjustments

+++ lib/skills/frozen_orb/FrozenOrbGraphicsComponent.cs
@@ -13,7 +13,7 @@ public class FrozenOrbGraphicsComponent
             new((int)frozenOrb.Position.X, (int)frozenOrb.Position.Y),
             _asset.Frames[0],
             Color.White,
-            (float)frozenOrb.Rotation,
+            frozenOrb.Rotation,
             new Vector2(_asset.Texture.Width / _asset.Frames.Count / 2, _asset.Texture.Height / 2),
             1f,
             SpriteEffects.None,
